Audio and other fixes
Fixed the radial menu bug where it would highlight over the text on the
upper right radial menu option.

Fixed the backwards lvl2 glass image

Added in audio sfx for glass, metal and wood breaking.  The sounds
currently loop a bajillion times for some reason when the game checks if
the slots are set to fail. Once that gets sorted out the sounds will
play randomly as the house burns down.